# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VBUY — Premium Ghanaian E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters and pagination
      parameters:
        - name: category
          in: query
          schema: { type: string }
          description: Category slug
        - name: brand
          in: query
          schema: { type: string }
          description: Brand slug
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular]
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get single product with variants and related products
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Product not found
  /products/{id}/related:
    get:
      operationId: listRelatedProducts
      tags: [products]
      summary: Get related products for a given product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductSummary"

  # ── Cart ─────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductImageAsset:
      type: object
      properties:
        primaryImage: { type: string }
        thumbnail:    { type: string }
        largePath:    { type: string }
        mediumPath:   { type: string }
        width:        { type: integer }
        height:       { type: integer }
        verified:     { type: boolean }
      required: [primaryImage, thumbnail, largePath, mediumPath]
    ProductSummary:
      type: object
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: string }
        basePrice:   { type: number }
        images:
          type: array
          items: { type: string }
        imageAssets:
          type: array
          items:
            $ref: "#/components/schemas/ProductImageAsset"
        categoryId:  { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        brandId:     { type: ["integer", "null"] }
        brandName:   { type: ["string", "null"] }
        inStock:     { type: boolean }
        variantCount: { type: integer }
      required: [id, name, description, basePrice, images, inStock, variantCount]
    ProductVariant:
      type: object
      properties:
        id:         { type: integer }
        name:       { type: string }
        sku:        { type: string }
        price:      { type: number }
        stockCount: { type: integer }
        attributes:
          type: object
          properties:
            color:   { type: ["string", "null"] }
            size:    { type: ["string", "null"] }
            storage: { type: ["string", "null"] }
            ram:     { type: ["string", "null"] }
            variant: { type: ["string", "null"] }
      required: [id, name, sku, price, stockCount]
    ProductDetail:
      type: object
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: string }
        basePrice:   { type: number }
        images:
          type: array
          items: { type: string }
        imageAssets:
          type: array
          items:
            $ref: "#/components/schemas/ProductImageAsset"
        categoryId:  { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        brandId:     { type: ["integer", "null"] }
        brandName:   { type: ["string", "null"] }
        inStock:     { type: boolean }
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
      required: [id, name, description, basePrice, images, inStock, variants]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductSummary"
        total:      { type: integer }
        page:       { type: integer }
        totalPages: { type: integer }
      required: [products, total, page, totalPages]
...